Update project structure and standards
