+++ composer.json
@@ -13,32 +13,32 @@
     "require": {
         "php": "^7.3",
         "ext-json": "*",
-        "keboola/csv": "^1.1",
+        "keboola/csv": "^2.2",
         "keboola/db-adapter-snowflake": "1.0.4",
-        "keboola/db-extractor-common": "^12.2",
-        "keboola/db-extractor-config": "^1.1",
-        "keboola/db-extractor-logger": "^1.0",
-        "keboola/db-extractor-table-format": "^1.3",
-        "keboola/php-datatypes": "^4.2",
-        "keboola/php-temp": "^1.0",
-        "keboola/retry": "^0.4.1",
-        "symfony/process": "^4.3"
+        "keboola/db-extractor-common": "^13.3",
+        "keboola/db-extractor-config": "^1.4",
+        "keboola/db-extractor-table-format": "^3.1",
+        "keboola/php-component": "^8.1",
+        "keboola/php-datatypes": "^4.9",
+        "keboola/php-temp": "^2.0",
+        "keboola/retry": "^0.5.0",
+        "symfony/process": "^5.2"
     },
     "require-dev": {
-        "codeclimate/php-test-reporter": "^0.4",
-        "keboola/coding-standard": "^8.0",
-        "keboola/storage-api-client": "^10.7",
-        "phpstan/phpstan-shim": "^0.11.15",
-        "phpunit/phpunit": "^8.3"
+        "keboola/coding-standard": "^9.0",
+        "keboola/datadir-tests": "^5.2",
+        "phpstan/phpstan": "^0.12.68",
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
-            "Keboola\\": "src/Keboola/"
+            "Keboola\\DbExtractor\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Keboola\\DbExtractor\\Tests\\": "tests/"
+            "Keboola\\DbExtractor\\Tests\\": "tests/phpunit/",
+            "Keboola\\DbExtractor\\TraitTests\\": "tests/traits/"
         }
     },
     "scripts": {
