+++ docker-compose.yml
@@ -18,3 +18,4 @@ services:
     <<: *base
     volumes:
       - .:/code
+      - ./data:/data
