+++ phpunit.xml.dist
@@ -6,7 +6,7 @@
          convertNoticesToExceptions="true"
          convertWarningsToExceptions="true"
          processIsolation="false"
-         stopOnFailure="false"
+         stopOnFailure="true"
          bootstrap="vendor/autoload.php">
 
     <testsuite name="DB Extractor Snowflake Test Suite">
