--- src/Keboola/DbExtractor/Extractor/Snowflake.php
@@ -1,731 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Keboola\DbExtractor\Extractor;
-
-use Keboola\Csv\CsvFile;
-use Keboola\Datatype\Definition\Exception\InvalidLengthException;
-use Keboola\DbExtractor\DbRetryProxy;
-use Keboola\DbExtractor\Exception\UserException;
-use Keboola\DbExtractor\TableResultFormat\Table;
-use Keboola\DbExtractor\TableResultFormat\TableColumn;
-use Keboola\DbExtractorLogger\Logger;
-use Keboola\DbExtractor\Utils\AccountUrlParser;
-use Keboola\Datatype\Definition\GenericStorage as GenericDatatype;
-use Keboola\Datatype\Definition\Snowflake as SnowflakeDatatype;
-use Keboola\Datatype\Definition\Exception\InvalidTypeException;
-use Keboola\SnowflakeDbAdapter\Connection;
-use Keboola\SnowflakeDbAdapter\Exception\CannotAccessObjectException;
-use Keboola\SnowflakeDbAdapter\QueryBuilder;
-use Keboola\Temp\Temp;
-use Symfony\Component\Process\Process;
-
-class Snowflake extends Extractor
-{
-    public const INCREMENT_TYPE_NUMERIC = 'numeric';
-    public const INCREMENT_TYPE_TIMESTAMP = 'timestamp';
-    public const INCREMENT_TYPE_DATE = 'date';
-    public const NUMERIC_BASE_TYPES = ['INTEGER', 'NUMERIC', 'FLOAT'];
-    public const SEMI_STRUCTURED_TYPES = ['VARIANT' , 'OBJECT', 'ARRAY'];
-
-    /** @var Connection */
-    protected $db;
-
-    /** @var \SplFileInfo */
-    private $snowSqlConfig;
-
-    /** @var string */
-    private $warehouse;
-
-    /** @var string */
-    private $database;
-
-    /** @var string */
-    private $schema;
-
-    /** @var string */
-    private $user;
-
-    /** @var Temp */
-    private $temp;
-
-    public function __construct(array $parameters, array $state, Logger $logger)
-    {
-        $this->temp = new Temp('ex-snowflake');
-
-        parent::__construct($parameters, $state, $logger);
-    }
-
-    public function createConnection(array $dbParams): Connection
-    {
-        $dbParams['password'] = $dbParams['#password'];
-        $this->snowSqlConfig = $this->createSnowSqlConfig($dbParams);
-        unset($dbParams['#password']);
-        unset($dbParams['ssh']);
-
-        if (getenv('KBC_RUNID')) {
-            $dbParams['runId'] = getenv('KBC_RUNID');
-        }
-
-        $this->user = $dbParams['user'];
-
-        $this->database = $dbParams['database'];
-
-        if (!empty($dbParams['warehouse'])) {
-            $this->warehouse = $dbParams['warehouse'];
-        }
-
-        try {
-            $connection = new Connection($dbParams);
-            if (!empty($dbParams['schema'])) {
-                $this->schema = $dbParams['schema'];
-                $connection->query(sprintf('USE SCHEMA %s', QueryBuilder::quoteIdentifier($this->schema)));
-            }
-        } catch (CannotAccessObjectException $e) {
-            throw new UserException($e->getMessage(), 0, $e);
-        }
-
-        return $connection;
-    }
-
-    public function testConnection(): void
-    {
-        $this->db->query('SELECT current_date;');
-
-        $defaultWarehouse = $this->getUserDefaultWarehouse();
-        if (!$defaultWarehouse && !$this->warehouse) {
-            throw new UserException('Specify "warehouse" parameter');
-        }
-
-        $warehouse = (string) $defaultWarehouse;
-        if ($this->warehouse) {
-            $warehouse = (string) $this->warehouse;
-        }
-
-        try {
-            $this->db->query(sprintf(
-                'USE WAREHOUSE %s;',
-                QueryBuilder::quoteIdentifier($warehouse)
-            ));
-        } catch (\Throwable $e) {
-            if (preg_match('/Object does not exist/ui', $e->getMessage())) {
-                throw new UserException(sprintf('Invalid warehouse "%s" specified', $warehouse));
-            } else {
-                throw $e;
-            }
-        }
-    }
-
-    public function export(array $table): array
-    {
-        $outputTable = $table['outputTable'];
-
-        $this->logger->info('Exporting to ' . $outputTable);
-        $maxValue = null;
-        if (isset($table['table']) && isset($this->incrementalFetching)) {
-            $maxValue = $this->getMaxOfIncrementalFetchingColumn($table['table']);
-        }
-        $rowCount = $this->exportAndDownload($table);
-
-        $output = [
-            'outputTable' => $outputTable,
-            'rows' => $rowCount,
-        ];
-        // output state
-        if ($maxValue) {
-            $output['state']['lastFetchedRow'] = $maxValue;
-        }
-
-        return $output;
-    }
-
-    public function getMaxOfIncrementalFetchingColumn(array $table): ?string
-    {
-
-        if (isset($this->incrementalFetching['limit']) && $this->incrementalFetching['limit'] > 0) {
-            $fullsql = sprintf(
-                'SELECT %s FROM %s.%s',
-                QueryBuilder::quoteIdentifier($this->incrementalFetching['column']),
-                QueryBuilder::quoteIdentifier($table['schema']),
-                QueryBuilder::quoteIdentifier($table['tableName'])
-            );
-
-            $fullsql .= $this->createIncrementalAddon();
-
-            $fullsql .= sprintf(
-                ' LIMIT %s OFFSET %s',
-                1,
-                $this->incrementalFetching['limit'] - 1
-            );
-        } else {
-            $fullsql = sprintf(
-                'SELECT MAX(%s) as %s FROM %s.%s',
-                QueryBuilder::quoteIdentifier($this->incrementalFetching['column']),
-                QueryBuilder::quoteIdentifier($this->incrementalFetching['column']),
-                QueryBuilder::quoteIdentifier($table['schema']),
-                QueryBuilder::quoteIdentifier($table['tableName'])
-            );
-        }
-        $result = $this->runRetriableQuery($fullsql, 'Error fetching maximum value');
-        if (count($result) > 0) {
-            return $result[0][$this->incrementalFetching['column']];
-        }
-        return null;
-    }
-
-    private function getColumnInfo(string $query): array
-    {
-        // Create temporary view from the supplied query
-        $sql = sprintf(
-            'SELECT * FROM (%s) LIMIT 0;',
-            rtrim(trim($query), ';')
-        );
-
-        $this->runRetriableQuery(
-            $sql,
-            sprintf('DB query "%s" failed: ', rtrim(trim($query), ';'))
-        );
-
-        return $this->runRetriableQuery(
-            'DESC RESULT LAST_QUERY_ID()',
-            'Error fetching last query id'
-        );
-    }
-
-    private function exportAndDownload(array $table): int
-    {
-        if (!isset($table['query']) || $table['query'] === '') {
-            $query = $this->simpleQuery($table['table'], $table['columns']);
-            $columnInfo = $this->getColumnInfo($query);
-            $objectColumns = array_filter($columnInfo, function ($column): bool {
-                return in_array($column['type'], self::SEMI_STRUCTURED_TYPES);
-            });
-            if (!empty($objectColumns)) {
-                $query = $this->simpleQueryWithCasting($table['table'], $columnInfo);
-            }
-        } else {
-            $query = $table['query'];
-            $columnInfo = $this->getColumnInfo($query);
-        }
-
-        $tmpTableName = str_replace('.', '_', $table['outputTable']);
-        $this->cleanupTableStage($tmpTableName);
-
-        // copy into internal staging
-        $copyCommand = $this->generateCopyCommand($tmpTableName, $query);
-
-        $res = $this->runRetriableQuery(
-            $copyCommand,
-            sprintf('Copy Command: %s failed with message', $copyCommand)
-        );
-        $rowCount = (int) ($res[0]['rows_unloaded'] ?? 0);
-
-        if ($rowCount === 0) {
-            // query resulted in no rows, nothing left to do
-            return 0;
-        }
-
-        $this->logger->info('Downloading data from Snowflake');
-
-        $outputDataDir = $this->dataDir . '/out/tables/' . $tmpTableName . '.csv.gz';
-
-        @mkdir($outputDataDir, 0755, true);
-
-        $sql = [];
-        $sql[] = sprintf('USE DATABASE %s;', QueryBuilder::quoteIdentifier($this->database));
-
-        if ($this->schema) {
-            $sql[] = sprintf('USE SCHEMA %s;', QueryBuilder::quoteIdentifier($this->schema));
-        }
-
-        if ($this->warehouse) {
-            $sql[] = sprintf('USE WAREHOUSE %s;', QueryBuilder::quoteIdentifier($this->warehouse));
-        }
-
-        $sql[] = sprintf(
-            'GET @~/%s file://%s;',
-            $tmpTableName,
-            $outputDataDir
-        );
-
-        $snowSql = $this->temp->createTmpFile('snowsql.sql');
-        file_put_contents($snowSql->getPathname(), implode("\n", $sql));
-
-        $this->logger->debug(trim(implode("\n", $sql)));
-
-        // execute external
-        $command = sprintf(
-            'snowsql --noup --config %s -c downloader -f %s',
-            $this->snowSqlConfig,
-            $snowSql
-        );
-
-        $this->logger->debug(trim($command));
-
-        $process = Process::fromShellCommandline($command);
-        $process->setTimeout(null);
-        $process->run();
-
-        if (!$process->isSuccessful()) {
-            $this->logger->error(sprintf('Snowsql error, process output %s', $process->getOutput()));
-            $this->logger->error(sprintf('Snowsql error: %s', $process->getErrorOutput()));
-            throw new \Exception(sprintf(
-                'File download error occurred processing [%s]',
-                isset($table['name']) ? $table['name'] : $tmpTableName
-            ));
-        }
-
-        $csvFiles = $this->parseFiles($process->getOutput(), $outputDataDir);
-        $bytesDownloaded = 0;
-        foreach ($csvFiles as $csvFile) {
-            $bytesDownloaded += $csvFile->getSize();
-        }
-
-        file_put_contents(
-            $outputDataDir . '.manifest',
-            json_encode(
-                $this->createTableManifest(
-                    $table,
-                    array_map(
-                        function ($column): string {
-                            return $column['name'];
-                        },
-                        $columnInfo
-                    )
-                )
-            )
-        );
-
-        $this->logger->info(sprintf(
-            '%d files (%s) downloaded',
-            count($csvFiles),
-            $this->dataSizeFormatted((int) $bytesDownloaded)
-        ));
-
-        $this->cleanupTableStage($tmpTableName);
-
-        return $rowCount;
-    }
-
-    private function generateCopyCommand(string $stageTmpPath, string $query): string
-    {
-        $csvOptions = [];
-        $csvOptions[] = sprintf('FIELD_DELIMITER = %s', $this->quote(CsvFile::DEFAULT_DELIMITER));
-        $csvOptions[] = sprintf('FIELD_OPTIONALLY_ENCLOSED_BY = %s', $this->quote(CsvFile::DEFAULT_ENCLOSURE));
-        $csvOptions[] = sprintf('ESCAPE_UNENCLOSED_FIELD = %s', $this->quote('\\'));
-        $csvOptions[] = sprintf('COMPRESSION = %s', $this->quote('GZIP'));
-        $csvOptions[] = sprintf('NULL_IF=()');
-
-        return sprintf(
-            '
-            COPY INTO @~/%s/part
-            FROM (%s)
-
-            FILE_FORMAT = (TYPE=CSV %s)
-            HEADER = false
-            MAX_FILE_SIZE=50000000
-            OVERWRITE = TRUE
-            ;
-            ',
-            $stageTmpPath,
-            rtrim(trim($query), ';'),
-            implode(' ', $csvOptions)
-        );
-    }
-
-    private function createTableManifest(array $table, array $columns): array
-    {
-        $manifestData = [
-            'destination' => $table['outputTable'],
-            'delimiter' => CsvFile::DEFAULT_DELIMITER,
-            'enclosure' => CsvFile::DEFAULT_ENCLOSURE,
-            'primary_key' => $table['primaryKey'],
-            'incremental' => $table['incremental'],
-            'columns' => $columns,
-        ];
-
-        if (isset($table['table']) && isset($table['table']['tableName'])) {
-            $tables = $this->getTables([$table['table']]);
-            if (count($tables) > 0) {
-                $tableDetails = $tables[0];
-                $columnMetadata = [];
-                foreach ($tableDetails['columns'] as $column) {
-                    if (count($table['columns']) > 0 && !in_array($column['name'], $table['columns'])) {
-                        continue;
-                    }
-                    $datatypeKeys = ['length', 'nullable', 'default'];
-                    try {
-                        $datatype = new SnowflakeDatatype(
-                            $column['type'],
-                            array_intersect_key($column, array_flip($datatypeKeys))
-                        );
-                    } catch (InvalidTypeException $e) {
-                        if (!in_array($column['type'], self::SEMI_STRUCTURED_TYPES)) {
-                            $this->logger->warning(
-                                'Encountered irregular type: ' . $column['type'] . ' for culumn ' . $column['name']
-                            );
-                        }
-                        $datatype = new GenericDatatype(
-                            $column['type'],
-                            array_intersect_key($column, array_flip($datatypeKeys))
-                        );
-                    }
-
-                    $columnMetadata[$column['name']] = $datatype->toMetadata();
-                    $nonDatatypeKeys = array_diff_key($column, array_flip($datatypeKeys));
-                    foreach ($nonDatatypeKeys as $key => $value) {
-                        if ($key !== 'name') {
-                            $columnMetadata[$column['name']][] = [
-                                'key' => 'KBC.' . $key,
-                                'value'=> $value,
-                            ];
-                        }
-                    }
-                }
-                unset($tableDetails['columns']);
-                foreach ($tableDetails as $key => $value) {
-                    $manifestData['metadata'][] = [
-                        'key' => 'KBC.' . $key,
-                        'value' => $value,
-                    ];
-                }
-                $manifestData['column_metadata'] = $columnMetadata;
-            }
-        }
-
-        return $manifestData;
-    }
-
-    private function dataSizeFormatted(int $bytes): string
-    {
-        $base = log($bytes) / log(1024);
-        $suffixes = [' B', ' KB', ' MB', ' GB', ' TB'];
-        return round(pow(1024, $base - floor($base)), 2) . $suffixes[(int) floor($base)];
-    }
-
-    public function getTables(?array $tables = null): array
-    {
-        $sql = $this->schema ? 'SHOW TABLES IN SCHEMA' : 'SHOW TABLES IN DATABASE';
-        $arr = $this->runRetriableQuery($sql, 'Error fetching show schema');
-
-        $sql = $this->schema ? 'SHOW VIEWS IN SCHEMA' : 'SHOW VIEWS IN DATABASE';
-        $views = $this->runRetriableQuery($sql, 'Error fetching show views');
-        $arr = array_merge($arr, $views);
-
-        /** @var Table[] $tableDefs */
-        $tableDefs = [];
-        $sqlWhereElements = [];
-        foreach ($arr as $table) {
-            if ($this->shouldTableBeSkipped($table)) {
-                continue;
-            }
-            if (is_null($tables) || !(array_search($table['name'], array_column($tables, 'tableName')) === false)) {
-                $isView = array_key_exists('text', $table);
-                $fullTableId = $table['schema_name'] . '.' . $table['name'];
-                $tableFormat = new Table();
-                $tableFormat
-                    ->setSchema($table['schema_name'])
-                    ->setName($table['name'])
-                    ->setCatalog((isset($table['database_name'])) ? $table['database_name'] : null)
-                    ->setType($isView ? 'VIEW' : (isset($table['kind']) ? $table['kind'] : null));
-
-                if (isset($table['rows'])) {
-                    $tableFormat->setRowCount((int) $table['rows']);
-                }
-                $tableDefs[$fullTableId] = $tableFormat;
-                $sqlWhereElements[] = sprintf(
-                    '(table_schema = %s AND table_name = %s)',
-                    $this->quote($table['schema_name']),
-                    $this->quote($table['name'])
-                );
-            }
-        }
-
-        if (count($tableDefs) === 0) {
-            return [];
-        }
-
-        $sqlWhereClause = sprintf('WHERE %s', implode(' OR ', $sqlWhereElements));
-
-        $sql = 'SELECT * FROM information_schema.columns '
-            . $sqlWhereClause
-            . ' ORDER BY TABLE_SCHEMA, TABLE_NAME, ORDINAL_POSITION';
-
-        $columns = $this->runRetriableQuery($sql, 'Error fetching table columns');
-        foreach ($columns as $i => $column) {
-            $curTable = $column['TABLE_SCHEMA'] . '.' . $column['TABLE_NAME'];
-            $length = ($column['CHARACTER_MAXIMUM_LENGTH']) ? $column['CHARACTER_MAXIMUM_LENGTH'] : null;
-            if (is_null($length) && !is_null($column['NUMERIC_PRECISION'])) {
-                if (is_numeric($column['NUMERIC_SCALE'])) {
-                    $length = $column['NUMERIC_PRECISION'] . ',' . $column['NUMERIC_SCALE'];
-                } else {
-                    $length = $column['NUMERIC_PRECISION'];
-                }
-            }
-            $columnFormat = new TableColumn();
-            $columnFormat
-                ->setName($column['COLUMN_NAME'])
-                ->setDefault($column['COLUMN_DEFAULT'])
-                ->setLength($length)
-                ->setNullable((trim($column['IS_NULLABLE']) === 'NO') ? false : true)
-                ->setType($column['DATA_TYPE'])
-                ->setOrdinalPosition((int) $column['ORDINAL_POSITION']);
-
-            $tableDefs[$curTable]->addColumn($columnFormat);
-        }
-        array_walk($tableDefs, function (Table &$item): void {
-            $item = $item->getOutput();
-        });
-        return array_values($tableDefs);
-    }
-
-    private function shouldTableBeSkipped(array $table): bool
-    {
-        $isFromDifferentSchema = $this->schema && $table['schema_name'] !== $this->schema;
-        $isFromInformationSchema = $table['schema_name'] === 'INFORMATION_SCHEMA';
-        return $isFromDifferentSchema || $isFromInformationSchema;
-    }
-
-    public function validateIncrementalFetching(array $table, string $columnName, ?int $limit = null): void
-    {
-        $columns = $this->runRetriableQuery(
-            sprintf(
-                'SELECT * FROM INFORMATION_SCHEMA.COLUMNS as cols 
-                            WHERE TABLE_SCHEMA = %s AND TABLE_NAME = %s AND COLUMN_NAME = %s',
-                $this->quote($table['schema']),
-                $this->quote($table['tableName']),
-                $this->quote($columnName)
-            ),
-            'Error get column info'
-        );
-        if (count($columns) === 0) {
-            throw new UserException(
-                sprintf(
-                    'Column [%s] specified for incremental fetching was not found in the table',
-                    $columnName
-                )
-            );
-        }
-
-        try {
-            $datatype = new SnowflakeDatatype($columns[0]['DATA_TYPE']);
-            if (in_array($datatype->getBasetype(), self::NUMERIC_BASE_TYPES)) {
-                $this->incrementalFetching['column'] = $columnName;
-                $this->incrementalFetching['type'] = self::INCREMENT_TYPE_NUMERIC;
-            } else if ($datatype->getBasetype() === 'TIMESTAMP') {
-                $this->incrementalFetching['column'] = $columnName;
-                $this->incrementalFetching['type'] = self::INCREMENT_TYPE_TIMESTAMP;
-            } else if ($datatype->getBasetype() === 'DATE') {
-                $this->incrementalFetching['column'] = $columnName;
-                $this->incrementalFetching['type'] = self::INCREMENT_TYPE_DATE;
-            } else {
-                throw new UserException('invalid incremental fetching column type');
-            }
-        } catch (InvalidLengthException | UserException $exception) {
-            throw new UserException(
-                sprintf(
-                    'Column [%s] specified for incremental fetching is not a numeric, date or timestamp type column',
-                    $columnName
-                )
-            );
-        }
-
-        if ($limit) {
-            $this->incrementalFetching['limit'] = $limit;
-        }
-    }
-
-    public function simpleQuery(array $table, array $columns = array()): string
-    {
-        $incrementalAddon = null;
-        if ($this->incrementalFetching && isset($this->incrementalFetching['column'])) {
-            $incrementalAddon = $this->createIncrementalAddon();
-        }
-        if (count($columns) > 0) {
-            $query = sprintf(
-                'SELECT %s FROM %s.%s',
-                implode(', ', array_map(function ($column): string {
-                    return QueryBuilder::quoteIdentifier($column);
-                }, $columns)),
-                QueryBuilder::quoteIdentifier($table['schema']),
-                QueryBuilder::quoteIdentifier($table['tableName'])
-            );
-        } else {
-            $query = sprintf(
-                'SELECT * FROM %s.%s',
-                QueryBuilder::quoteIdentifier($table['schema']),
-                QueryBuilder::quoteIdentifier($table['tableName'])
-            );
-        }
-        if ($incrementalAddon) {
-            $query .= $incrementalAddon;
-        }
-        if (isset($this->incrementalFetching['limit'])) {
-            $query .= sprintf(
-                ' LIMIT %d',
-                $this->incrementalFetching['limit']
-            );
-        }
-        return $query;
-    }
-
-    private function createIncrementalAddon(): string
-    {
-        $incrementalAddon = '';
-        if (isset($this->state['lastFetchedRow'])) {
-            if ($this->incrementalFetching['type'] === self::INCREMENT_TYPE_NUMERIC) {
-                $lastFetchedRow = $this->state['lastFetchedRow'];
-            } else {
-                $lastFetchedRow = $this->quote((string) $this->state['lastFetchedRow']);
-            }
-            $incrementalAddon = sprintf(
-                ' WHERE %s >= %s',
-                QueryBuilder::quoteIdentifier($this->incrementalFetching['column']),
-                $lastFetchedRow
-            );
-        }
-        $incrementalAddon .= sprintf(
-            ' ORDER BY %s',
-            QueryBuilder::quoteIdentifier($this->incrementalFetching['column'])
-        );
-        return $incrementalAddon;
-    }
-
-    private function simpleQueryWithCasting(array $table, array $columnInfo): string
-    {
-        return sprintf(
-            'SELECT %s FROM %s.%s',
-            implode(', ', array_map(function ($column): string {
-                if (in_array($column['type'], self::SEMI_STRUCTURED_TYPES)) {
-                    return sprintf(
-                        'CAST(%s AS TEXT) AS %s',
-                        QueryBuilder::quoteIdentifier($column['name']),
-                        QueryBuilder::quoteIdentifier($column['name'])
-                    );
-                }
-                return QueryBuilder::quoteIdentifier($column['name']);
-            }, $columnInfo)),
-            QueryBuilder::quoteIdentifier($table['schema']),
-            QueryBuilder::quoteIdentifier($table['tableName'])
-        );
-    }
-
-    private function parseFiles(string $output, string $path): array
-    {
-        $files = [];
-        $lines = explode("\n", $output);
-
-        $lines = array_map(
-            function ($item): array {
-                $item = trim($item, '|');
-                return array_map('trim', explode('|', $item));
-            },
-            array_filter(
-                $lines,
-                function ($item): bool {
-                    $item = trim($item);
-                    return preg_match('/^\|.+\|$/ui', $item) && preg_match('/([a-z0-9\_\-\.]+\.gz)/ui', $item);
-                }
-            )
-        );
-
-        foreach ($lines as $line) {
-            if (!preg_match('/^downloaded$/ui', $line[2])) {
-                throw new \Exception(sprintf(
-                    'Cannot download file: %s Status: %s Size: %s Message: %s',
-                    $line[0],
-                    $line[2],
-                    $line[1],
-                    $line[3]
-                ));
-            }
-
-            $file = new \SplFileInfo($path . '/' . $line[0]);
-            if ($file->isFile()) {
-                $files[] = $file;
-            } else {
-                throw new \Exception('Missing file: ' . $line[0]);
-            }
-        }
-
-        return $files;
-    }
-
-    private function quote(string $value): string
-    {
-        return "'" . addslashes($value) . "'";
-    }
-
-    private function createSnowSqlConfig(array $dbParams): \SplFileInfo
-    {
-        $cliConfig[] = '';
-        $cliConfig[] = '[options]';
-        $cliConfig[] = 'exit_on_error = true';
-        $cliConfig[] = '';
-        $cliConfig[] = '[connections.downloader]';
-        $cliConfig[] = sprintf('accountname = "%s"', AccountUrlParser::parse($dbParams['host']));
-        $cliConfig[] = sprintf('username = "%s"', $dbParams['user']);
-        $cliConfig[] = sprintf('password = "%s"', $dbParams['#password']);
-        $cliConfig[] = sprintf('dbname = "%s"', $dbParams['database']);
-
-        if (isset($dbParams['warehouse'])) {
-            $cliConfig[] = sprintf('warehousename = "%s"', $dbParams['warehouse']);
-        }
-
-        if (isset($dbParams['schema'])) {
-            $cliConfig[] = sprintf('schemaname = "%s"', $dbParams['schema']);
-        }
-
-        $file = $this->temp->createFile('snowsql.config');
-        file_put_contents($file->getPathname(), implode("\n", $cliConfig));
-
-        return $file;
-    }
-
-    private function getUserDefaultWarehouse(): ?string
-    {
-        $sql = sprintf(
-            'DESC USER %s;',
-            QueryBuilder::quoteIdentifier($this->user)
-        );
-
-        $config = $this->runRetriableQuery($sql, 'Error get user config');
-
-        foreach ($config as $item) {
-            if ($item['property'] === 'DEFAULT_WAREHOUSE') {
-                return $item['value'] === 'null' ? null : $item['value'];
-            }
-        }
-
-        return null;
-    }
-
-    private function runRetriableQuery(string $query, string $errorMessage = '', string $type = 'fetchAll'): array
-    {
-        $retryProxy = new DbRetryProxy(
-            $this->logger,
-            DbRetryProxy::DEFAULT_MAX_TRIES,
-            [\PDOException::class, \ErrorException::class, \Throwable::class]
-        );
-        return $retryProxy->call(function () use ($query, $type, $errorMessage): array {
-            try {
-                /** @var array $result */
-                $result = (array) $this->db->{$type}($query);
-                return $result;
-            } catch (\Throwable $e) {
-                throw new UserException(
-                    $errorMessage . ': ' . $e->getMessage(),
-                    0,
-                    $e
-                );
-            }
-        });
-    }
-
-    private function cleanupTableStage(string $tmpTableName): void
-    {
-        $sql = sprintf('REMOVE @~/%s;', $tmpTableName);
-        $this->runRetriableQuery($sql, 'Query execution error', 'query');
-    }
-}
