--- tests/data/connectionAction/config.row.template.json
@@ -1,15 +1,0 @@
-{
-    "action": "testConnection",
-    "parameters": {
-        "extractor_class": "snowflake",
-        "id": 2,
-        "name": "escaping",
-        "query": "SELECT * FROM \"escaping\" ;\n",
-        "outputTable": "in.c-main.escaping",
-        "incremental": false,
-        "primaryKey": [
-            "col1"
-        ],
-        "enabled": true
-    }
-}
