--- tests/data/getTablesAction/config.template.json
@@ -1,38 +1,0 @@
-{
-    "action": "getTables",
-    "parameters": {
-        "tables": [
-            {
-                "id": 1,
-                "name": "sales",
-                "query": "SELECT * FROM \"sales\"",
-                "outputTable": "in.c-main.sales",
-                "incremental": false,
-                "primaryKey": null,
-                "enabled": true
-            },
-            {
-                "id": 2,
-                "name": "escaping",
-                "query": "SELECT * FROM \"escaping\"",
-                "outputTable": "in.c-main.escaping",
-                "incremental": false,
-                "primaryKey": [
-                    "col1"
-                ],
-                "enabled": true
-            },
-            {
-                "id": 3,
-                "name": "somedata",
-                "query": "SELECT * FROM \"somedata\"",
-                "outputTable": "in.c-main.somedata",
-                "incremental": true,
-                "primaryKey": [
-                    "somedataId"
-                ],
-                "enabled": false
-            }
-        ]
-    }
-}
