--- tests/data/runAction/config.row.template.json
@@ -1,14 +1,0 @@
-{
-    "parameters": {
-        "extractor_class": "snowflake",
-        "id": 2,
-        "name": "escaping",
-        "query": "SELECT * FROM \"escaping\" ;\n",
-        "outputTable": "in.c-main.escaping",
-        "incremental": false,
-        "primaryKey": [
-            "col1"
-        ],
-        "enabled": true
-    }
-}
