--- tests/data/runAction/config.rowTable.template.json
@@ -1,17 +1,0 @@
-{
-    "parameters": {
-        "extractor_class": "snowflake",
-        "id": 2,
-        "table": {
-            "schema": null,
-            "tableName": "escaping"
-        },
-        "name": "escaping",
-        "outputTable": "in.c-main.escaping",
-        "incremental": false,
-        "primaryKey": [
-            "col1"
-        ],
-        "enabled": true
-    }
-}
