--- tests/data/snowflake/config.json
@@ -1,44 +1,0 @@
-{
-  "parameters": {
-    "tables": [
-    {
-      "id": 1,
-      "name": "sales",
-      "query": "SELECT * FROM \"sales\" WHERE \"categorygroup\" NOT LIKE '%stupidsprintf'",
-      "outputTable": "in.c-main.sales",
-      "incremental": false,
-      "primaryKey": null,
-      "enabled": true
-    },
-    {
-      "id": 2,
-      "name": "escaping",
-      "query": "SELECT * FROM \"escaping\" ;\n",
-      "outputTable": "in.c-main.escaping",
-      "incremental": false,
-      "primaryKey": [
-        "col1"
-      ],
-      "enabled": true
-    },
-    {
-      "id": 3,
-      "name": "tableColumns",
-      "outputTable": "in.c-main.tableColumns",
-      "incremental": false,
-      "primaryKey": null,
-      "enabled": true,
-      "table": {
-        "schema": null,
-        "tableName": "types"
-      },
-      "columns": [
-        "character",
-        "integer",
-        "decimal",
-        "date"
-      ]
-    }
-    ]
-  }
-}
