--- tests/data/snowflake/configRow.json
@@ -1,13 +1,0 @@
-{
-  "parameters": {
-    "id": 1,
-    "name": "escaping",
-    "query": "SELECT * FROM \"escaping\" ;\n",
-    "outputTable": "in.c-main.escaping",
-    "incremental": false,
-    "primaryKey": [
-      "col1"
-    ],
-    "enabled": true
-  }
-}
