--- tests/data/snowflake/escaping.csv
@@ -1,11 +1,0 @@
-"col1","col2"
-"line with enclosure","second column"
-"column with enclosure "", and comma inside text","second column enclosure in text """
-"columns with
-new line","columns with 	tab"
-"column with backslash \ inside","column with backslash and enclosure \"""
-"column with \n \t \\","second col"
-"unicode characters","ľščťžýáíéúäôň"
-"first","something with
-
-double new line"
