--- tests/data/snowflake/expected-semi-structured.csv
@@ -1,1 +1,0 @@
-"{""a"":1,""b"":""BBBB""}","{""a"":1,""b"":""BBBB""}","[10,20,30]"
