--- tests/data/snowflake/tableColumns.csv
@@ -1,101 +1,0 @@
-"usergender","usercity","usersentiment","zipcode"
-"usergender","usercity","","zipcode"
-"Female","Mize","-1","39153"
-"Male","The Lakes","1","89124"
-"Male","Baldwin","1","21020"
-"Female","Archbald","1","18501"
-"Male","Berea","0","44127"
-"Male","Baldwin","0","21219"
-"Male","Phoenix","1","85083"
-"Female","Martinsburg","-1","25428"
-"Male","Los Angeles","0","91328"
-"Female","Bolton","1","28455"
-"Female","Cranford","-1","07974"
-"Female","Troy","1","48359"
-"Female","East Mansfield","1","02764"
-"Male","Bixby","0","74156"
-"Female","Buena Park","0","92806"
-"Male","Annapolis Junction","0","21029"
-"Male","Bellarthur","-1","27837"
-"Male","Longmont","1","80615"
-"Female","Allenton","0","63126"
-"Female","Red Creek","-1","14551"
-"Male","Knightstown","-1","47351"
-"Female","Kimball","0","56372"
-"Female","Brighton","1","48836"
-"Female","Coloma","-1","96155"
-"Male","Basehor","0","66027"
-"Male","Alma Center","0","54643"
-"Female","Monroe","0","71291"
-"Male","Arlington Heights","0","60630"
-"Male","Englewood","1","34231"
-"Female","Allen Park","1","48210"
-"Female","Los Angeles","1","91328"
-"Female","Los Angeles","-1","91386"
-"Male","Charlottesville","1","22940"
-"Female","Los Angeles","0","92119"
-"Female","Bay Pines","0","33755"
-"Female","Cornelius","1","28289"
-"Male","Clifton","1","07506"
-"Female","Corona","-1","92518"
-"Male","Edgard","-1","70049"
-"Female","Durham","0","18951"
-"Male","Copperas Cove","-1","76558"
-"Female","Brookville","0","45403"
-"Male","Los Angeles","0","90034"
-"Male","Arkansas City","0","72379"
-"Female","Edgewood","1","52042"
-"Male","Eastwood","-1","40285"
-"Female","Catoosa","1","74015"
-"Female","Roy","1","84402"
-"Female","Los Angeles","-1","91507"
-"Female","Lightfoot","-1","23694"
-"Female","Butler","0","16046"
-"Male","Ludlow","-1","61880"
-"Female","Los Angeles","0","92038"
-"Female","Cornelius","0","28217"
-"Male","Allenton","0","63132"
-"Female","Hattiesburg","0","39402"
-"Male","Nantucket","1","02564"
-"Female","Dania","0","33309"
-"Female","Jamaica","-1","14215"
-"Female","Bend","0","97707"
-"Female","Dania","-1","33326"
-"Male","Dania","-1","33339"
-"Female","West Palm Beach","0","33454"
-"Male","Batesburg","-1","29169"
-"Female","Atlas","0","48458"
-"Female","Andale","1","67202"
-"Female","Discovery Bay","-1","94507"
-"Female","Monroe","0","71294"
-"Female","Bass","-1","72628"
-"Male","Baldwin City","-1","66046"
-"Female","Centreville","-1","22109"
-"Male","Allenton","-1","63127"
-"Female","Okemos","1","48911"
-"Female","Denver","0","80219"
-"Male","Gainesville","1","32606"
-"Female","Troy","0","48086"
-"Female","Bay Minette","-1","36550"
-"Female","Arab","1","35957"
-"Male","Hialeah","0","33128"
-"Female","Sacramento","1","95887"
-"Male","McHenry","-1","62034"
-"Female","Philadelphia","-1","19128"
-"Female","Bradley","0","93923"
-"Female","Gering","-1","69358"
-"Female","Los Angeles","0","90223"
-"Female","Amboy","1","98686"
-"Female","Barberville","1","32724"
-"Male","Carrollton","0","76258"
-"Female","Saint Charles","-1","63303"
-"Female","Bairdford","-1","15270"
-"Male","Glen Oaks","0","11447"
-"Female","Addison","1","75015"
-"Male","Bible School Park","1","13777"
-"Male","Jersey","1","30025"
-"Male","Hanna","-1","46346"
-"Male","Saint Paul","1","55375"
-"Female","Glen Oaks","1","11366"
-"Male","Alexandria","-1","17260"
-"Male","Los Angeles","0","92124"
