--- tests/data/snowflake/types.csv
@@ -1,5 +1,0 @@
-"character", "integer", "decimal", "date"
-"abcdef",12,22.40,2017-08-05
-"abcdefg",13,22.42,2017-08-06
-"abcdefgh",32,22.41,2017-08-07
-"abcdefghi",22,22.44,2017-08-08
