+++ tests/traits/QuoteIdentifierTrait.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\DbExtractor\TraitTests;
+
+trait QuoteIdentifierTrait
+{
+
+    public function quoteIdentifier(string $str): string
+    {
+        return sprintf('"%s"', $str);
+    }
+}
